Рефактор утилитных функций
