+++ src/mock/profile.js
@@ -1,4 +1,4 @@
-import {getRandomIntegerNumber} from '../utils.js';
+import {getRandomIntegerNumber} from '../utils/common';
 
 export const generateUser = () => ({
   filmsCount: getRandomIntegerNumber(0, 100)
